Calculo de costo y distancia funcional

+++ src/main/java/FormatoBase/proyectoJWT/model/dto/OrderCostResponse.java
@@ -11,6 +11,6 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderCostResponse implements Serializable {
-    private Integer distance;
+    private BigDecimal distance;
     private BigDecimal totalCost;
 }
